* added function 'build_element': function isn't done yet, made to create xml elements for each new data entry calculated. Current code: 'pass'
* added function 'highest_priority':
function fetches the highest integer found when parsing the xml source file for
element property called 'priority' which is a child element in every 'search' element
being parsed. This needs to be numeric for each new element, thus, needs to be fetched
to know where to start building from with new elements.

* added function: 'get_new_data':
function takes file with new data as argument, and creates a set() to be returned.

* added function 'get_old_data':
function takes root from get_xml_root and adds every old text content for elements
named 'orgName' which are attributes in Active Directory of interest.
These are added to set() and returned.

* added function 'get_xml_root':
function gets the root for the .xml file and returns it.
made to its own function for easily create a root object in separate functions,
for code re-use.